LayoutTests/http/tests/inspector/network/network-request-revision-content.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script>
function loadScript()
{
    var js = document.createElement("script");
    js.src = "resources/script.js";
    document.head.appendChild(js);
}

function test()
{
    InspectorTest.addConsoleSniffer(step2, true);
    InspectorTest.evaluateInPage("loadScript()");

    var resource;
    function step2(msg)
    {
        var request = WebInspector.panels.network.requests[WebInspector.panels.network.requests.length - 1];
        resource = WebInspector.resourceForURL(request.url);
        resource.addRevision("");
        resource.history[0].requestContent(step3);
    }

    function step3(content)
    {
        InspectorTest.addResult(resource.url);
        InspectorTest.addResult(content);
        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests how resource revision requests content if it was not yet loaded for resource itself.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=63631">Bug 63631</a>
</body>
</html>
